pipeline/requirements.txt
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
